register operations under generic sat policy instead
