+++ src/schedlib/policies/__init__.py
@@ -1,2 +1,3 @@
+from .flex import FlexPolicy
 from .sat import SATPolicy
 from .satp1 import SATP1Policy
